initial commit
Why:

Development should be fun and productive but there are risks which
jeopardize the fun and productivity of development. This is true in
software development and all aspects of life. PyDX tries to tame
these risks in the context of the Python Developer eXperience by
documenting and sharing relevant risks, solutions, and their
rewards.

This change addresses the need by:

* initializes knowledge repository