feat(git): integrate pre-commit Git hooks framework
Why
===

"No code is the best way to write secure and reliable applications.
 Write nothing; deploy nowhere." -- Kelsey Hightower

Until the No Code skill level is achieved Git hooks are a great way
to ensure the code that is written is lint free. However, managing
Git hooks can be as tedious as the debates linting may provoke. It's
best to just adopt a Git hooks framework such as pre-commit or Husky
and define a standard set hook integrations.

This change addresses the need by
=================================

* integrates the pre-commit Git hooks framework
* configures hooks' default stage
* configures subsequent hooks to not run should one fail
* excludes .git folder from checks
* integrates 1st hook which excludes files over 1MB from commits
* adds a GNU Make rule for running pre-commit checks as needed
* begins GNU Make rules for setting up the local development environment
  including the installation of pre-commit and configured hooks
* begins use of restructured text in commit messages